+++ servers/mcp-neo4j-cypher/CHANGELOG.md
@@ -3,8 +3,12 @@
 ### Fixed
 
 ### Changed
+* Update error handling in read and write tools
 
 ### Added
+* Add .dxt file for Cypher MCP server
+* Add .dxt file generation to Cypher MCP Publish GitHub action
+* Add error indicator to tool results in the `CallToolResult` object
 
 ## v0.2.4